+++ packages/terra-collapsible-menu-view/tests/nightwatch/CollapsibleMenuViewTests.jsx
@@ -7,6 +7,7 @@ const CollapsibleMenuViewTests = () => (
   <div>
     <ul>
       <li><Link to="/tests/collapsible-menu-view-tests/default">CollapsibleMenuView - Default</Link></li>
+      <li><Link to="/tests/collapsible-menu-view-tests/first-item-hidden">CollapsibleMenuView - First Hidden Item Is Icon Only</Link></li>
     </ul>
   </div>
 );