+++ packages/terra-form-select/tests/nightwatch/select/DisabledSelect.jsx
@@ -1,0 +1,28 @@
+import React from 'react';
+import Base from 'terra-base';
+import Select from '../../../lib/Select';
+
+const locale = document.getElementsByTagName('html')[0].getAttribute('lang');
+
+const DisabledSelect = () => (
+  <Base locale={locale}>
+    <p>Disabled select with a preselected value of kittens.</p>
+    <form>
+      <Select
+        name="zibby"
+        disabled
+        id="disabledSelect"
+      >
+        <Select.Option value="puppies" display="Puppies" key="puppies" />
+        <Select.Option value="kittens" display="Kittens" key="kittens" isSelected />
+        <Select.Option value="snappers" display="Snappers" key="snappers" />
+        <Select.Option value="bumblers" display="Bumblers" disabled key="bumblers" />
+        <Select.Option value="joeys" display="Joeys" key="joeys" />
+        <Select.Option value="micros" display="Microprocessors" disabled key="miros" />
+      </Select>
+    </form>
+    <br />
+  </Base>
+);
+
+export default DisabledSelect;