+++ packages/terra-search-field/src/SearchField.jsx
@@ -9,11 +9,30 @@ import styles from './SearchField.scss';
 
 const cx = classNames.bind(styles);
 
+const KEYCODES = {
+  ENTER: 13,
+};
+
 const propTypes = {
   /**
-   * Placeholder text to show while the search field is empty.
+   * The defaultValue of the search field. Use this to create an uncontrolled search field.
    */
-  placeholder: PropTypes.string,
+  defaultValue: PropTypes.string,
+
+  /**
+   * When true, will disable the auto-search.
+   */
+  disableAutoSearch: PropTypes.bool,
+
+  /**
+   * Whether or not the field should display as a block.
+   */
+  isBlock: PropTypes.bool,
+
+  /**
+   * When true, will disable the field.
+   */
+  isDisabled: PropTypes.bool,
 
   /**
    * The minimum number of characters to perform a search.
@@ -21,9 +40,19 @@ const propTypes = {
   minimumSearchTextLength: PropTypes.number,
 
   /**
-   * How long the component should wait (in milliseconds) after input before performing an automatic search.
+   * Placeholder text to show while the search field is empty.
    */
-  searchDelay: PropTypes.number,
+  placeholder: PropTypes.string,
+
+  /**
+   * Function to trigger when user changes the input value. Provide a function to create a controlled input.
+   */
+  onChange: PropTypes.func,
+
+  /**
+   * A callback to indicate an invalid search. Sends parameter {String} searchText.
+   */
+  onInvalidSearch: PropTypes.func,
 
   /**
    * A callback to perform search. Sends parameter {String} searchText.
@@ -31,15 +60,25 @@ const propTypes = {
   onSearch: PropTypes.func,
 
   /**
-   * A callback to indicate an invalid search. Sends parameter {String} searchText.
+   * How long the component should wait (in milliseconds) after input before performing an automatic search.
    */
-  onInvalidSearch: PropTypes.func,
+  searchDelay: PropTypes.number,
+
+  /**
+   * The value of search field.  Use this to create a controlled search field.
+   */
+  value: PropTypes.string,
 };
 
 const defaultProps = {
-  placeholder: '',
+  defaultValue: undefined,
+  disableAutoSearch: false,
+  isBlock: false,
+  isDisabled: false,
   minimumSearchTextLength: 2,
+  placeholder: '',
   searchDelay: 250,
+  value: undefined,
 };
 
 class SearchField extends React.Component {
@@ -49,11 +88,16 @@ class SearchField extends React.Component {
 
     this.handleTextChange = this.handleTextChange.bind(this);
     this.handleSearch = this.handleSearch.bind(this);
+    this.handleKeyDown = this.handleKeyDown.bind(this);
+    this.updateSearchText = this.updateSearchText.bind(this);
 
     this.searchTimeout = null;
-    this.state = {
-      searchText: '',
-    };
+    this.state = { searchText: this.props.defaultValue };
+  }
+
+  componentDidUpdate() {
+    // if consumer updates the value prop with onChange, need to update state to match
+    this.updateSearchText(this.props.value);
   }
 
   componentWillUnmount() {
@@ -61,10 +105,27 @@ class SearchField extends React.Component {
   }
 
   handleTextChange(event) {
-    this.setState({ searchText: event.target.value });
+    const textValue = event.target.value;
+    this.updateSearchText(textValue);
 
-    if (!this.searchTimeout) {
+    if (this.props.onChange) {
+      this.props.onChange(event, textValue);
+    }
+
+    if (!this.searchTimeout && !this.props.disableAutoSearch) {
       this.searchTimeout = setTimeout(this.handleSearch, this.props.searchDelay);
+    }
+  }
+
+  updateSearchText(searchText) {
+    if (searchText !== undefined && searchText !== this.state.searchText) {
+      this.setState({ searchText });
+    }
+  }
+
+  handleKeyDown(event) {
+    if (event.nativeEvent.keyCode === KEYCODES.ENTER) {
+      this.handleSearch();
     }
   }
 
@@ -86,11 +147,32 @@ class SearchField extends React.Component {
   }
 
   render() {
-    const { placeholder, searchDelay, minimumSearchTextLength, onSearch, onInvalidSearch, ...customProps } = this.props;
+    const {
+      defaultValue,
+      disableAutoSearch,
+      isBlock,
+      isDisabled,
+      minimumSearchTextLength,
+      placeholder,
+      searchDelay,
+      onChange,
+      onInvalidSearch,
+      onSearch,
+      value,
+      ...customProps
+    } = this.props;
     const searchFieldClassNames = cx([
       'searchfield',
+      { block: isBlock },
       customProps.className,
     ]);
+
+    const additionalInputAttributes = {};
+    if (value !== undefined) {
+      additionalInputAttributes.value = value;
+    } else {
+      additionalInputAttributes.defaultValue = defaultValue;
+    }
 
     return (
       <div {...customProps} className={searchFieldClassNames}>
@@ -98,13 +180,17 @@ class SearchField extends React.Component {
           className={cx('input')}
           type="search"
           placeholder={placeholder}
-          value={this.state.searchText}
           onChange={this.handleTextChange}
+          disabled={isDisabled}
+          aria-disabled={isDisabled}
+          onKeyDown={this.handleKeyDown}
+          {...additionalInputAttributes}
         />
         <Button
           className={cx('button')}
           onClick={this.handleSearch}
           isCompact
+          isDisabled={isDisabled}
         >
           <IconSearch />
         </Button>