+++ packages/terra-site/src/examples/search-field/SearchFieldDisableAutoSearch.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import SearchFieldExampleTemplate from './SearchFieldExampleTemplate';
+
+const SearchFieldDisableAutoSearch = () => (
+  <SearchFieldExampleTemplate
+    disableAutoSearch
+  />
+);
+
+export default SearchFieldDisableAutoSearch;