+++ packages/terra-base/CHANGELOG.md
@@ -4,6 +4,16 @@ Changelog
 Unreleased
 ----------
 
+2.8.0 - (November 28, 2017)
+------------------
+### Changed
+* componentWillReceiveProps will now compare locale instead of props.
+
+2.7.0 - (November 16, 2017)
+------------------
+### Changed
+* Minor version bump
+
 2.6.0 - (October 6, 2017)
 ------------------
 ### Changed