+++ packages/terra-divider/CHANGELOG.md
@@ -4,6 +4,16 @@ ChangeLog
 Unreleased
 -----------------
 
+1.2.0 - (November 28, 2017)
+------------------
+### Changed
+* Minor version bump
+
+1.1.0 - (November 16, 2017)
+------------------
+### Changed
+* Minor version bump
+
 1.0.0 - (October 31, 2017)
 ------------------
 Initial stable release