+++ packages/terra-embedded-content-consumer/CHANGELOG.md
@@ -4,6 +4,16 @@ ChangeLog
 Unreleased
 -----------------
 
+1.9.0 - (November 28, 2017)
+------------------
+### Changed
+* Minor version bump
+
+1.8.0 - (November 16, 2017)
+------------------
+### Changed
+* Minor version bump
+
 1.7.0 - (October 31, 2017)
 ------------------
 ### Changed