+++ packages/terra-form-field/CHANGELOG.md
@@ -1,6 +1,14 @@
 ChangeLog
 =========
 
-Unreleased 
+Unreleased
 -----------------
+
+1.1.0 - (November 28, 2017)
+------------------
+### Changed
+* Minor version bump
+
+1.0.0 - (November 16, 2017)
+------------------
 Initial stable release