+++ packages/terra-grid/CHANGELOG.md
@@ -4,6 +4,16 @@ Changelog
 Unreleased
 ----------
 
+3.17.0 - (November 28, 2017)
+------------------
+### Changed
+* Minor version bump
+
+3.16.0 - (November 16, 2017)
+------------------
+### Changed
+* Minor version bump
+
 3.15.0 - (October 6, 2017)
 ------------------
 ### Changed