+++ packages/terra-i18n/CHANGELOG.md
@@ -4,6 +4,16 @@ Changelog
 Unreleased
 ----------
 
+1.12.0 - (November 28, 2017)
+------------------
+### Changed
+* Minor version bump
+
+1.11.0 - (November 16, 2017)
+------------------
+### Changed
+* Minor version bump
+
 1.10.0 - (October 6, 2017)
 ------------------
 ### Changed