+++ packages/terra-list/CHANGELOG.md
@@ -3,6 +3,14 @@ Changelog
 
 Unreleased
 ----------
+
+1.17.0 - (November 28, 2017)
+------------------
+### Changed
+* Minor version bump
+
+1.16.0 - (November 16, 2017)
+------------------
 ### Fixed
 * Chevron positioning