+++ packages/terra-search-field/CHANGELOG.md
@@ -4,6 +4,18 @@ Changelog
 Unreleased
 ----------
 
+1.17.0 - (November 28, 2017)
+------------------
+### Changed
+* Add ability to default value and onChange function handler
+* Add ability to disable and stretch to fill its container
+* Add ability to disable the auto searching and trigger searching via ENTER key
+
+1.16.0 - (November 16, 2017)
+------------------
+### Changed
+* Minor version bump
+
 1.15.0 - (November 7, 2017)
 ------------------
 ### Changed