+++ packages/terra-site/CHANGELOG.md
@@ -3,6 +3,16 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Place site header background on a parent div, instead of collapsible.
+
+1.19.0 - (November 28, 2017)
+------------------
+### Changed
+* Minor version bump
+
+1.18.0 - (November 16, 2017)
+------------------
 ### Added
 * Add terra-form-textarea components
 * Add terra-form-field components