+++ packages/terra-time-input/CHANGELOG.md
@@ -3,6 +3,14 @@ Changelog
 
 Unreleased
 ----------
+
+1.15.0 - (November 28, 2017)
+------------------
+### Changed
+* Removed onInputFocus prop and pass it down as a custom prop.
+
+1.14.0 - (November 16, 2017)
+------------------
 ### Added
 * Added onInputFocus prop.